+++ fragments/arguments.sh
@@ -43,12 +43,14 @@ if [[ $label == "version" ]]; then
     exit 0
 fi
 
-printlog "################## Start Installomator v. $VERSION"
-printlog "################## $label"
+printlog "################## Start Installomator v. $VERSION, date $VERSIONDATE" REQ
+printlog "################## Version: $VERSION" INFO
+printlog "################## Date: $VERSIONDATE" INFO
+printlog "################## $label" INFO
 
 # Check for DEBUG mode
 if [[ $DEBUG -gt 0 ]]; then
-    printlog "DEBUG mode $DEBUG enabled."
+    printlog "DEBUG mode $DEBUG enabled." DEBUG
 fi
 
 # How we get version number from app